GH-15: endret litt struktur i koden
