+++ src/components/App.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Customer from "./Customer";
 import Transactions from "./Transactions";
-import { H2, Section, Tabs } from "@dnb/eufemia";
+import { Tabs } from "@dnb/eufemia";
 
 export default function App() {
   return (
@@ -11,15 +11,11 @@ export default function App() {
       content_style="black-3"
       style={{ marginLeft: "1rem", marginRight: "1rem" }}
     >
-      <Tabs.Content title="Kunde">
+      <Tabs.Content title="Privatkunde">
         <Customer />
       </Tabs.Content>
-      <Tabs.Content title="Transaksjoner">
-        <Section spacing top bottom style_type="white">
-          <H2 top={0} bottom>
-            <Transactions />
-          </H2>
-        </Section>
+      <Tabs.Content title="DNB">
+        <Transactions />
       </Tabs.Content>
     </Tabs>
   );
