+++ src/index.tsx
@@ -4,7 +4,7 @@ import { StrictMode } from "react";
 import "@dnb/eufemia/style";
 
 import App from "./components/App";
-import Footer from "./components/Footer";
+import Footer from "./div/Footer";
 
 ReactDOM.render(
   <StrictMode>
